+++ matrix-sdk-android/src/main/java/fr/gouv/tchap/android/sdk/internal/services/threepidplatformdiscover/TchapThreePidPlatformDiscoverService.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Matrix.org Foundation C.I.C.
+ * Copyright (c) 2021 New Vector Ltd
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package fr.gouv.tchap.android.sdk.internal.session.identity
+package fr.gouv.tchap.android.sdk.internal.services.threepidplatformdiscover
 
 import dagger.Lazy
-import fr.gouv.tchap.android.sdk.api.session.identity.IdentityService
-import fr.gouv.tchap.sdk.internal.session.identity.model.Platform
+import fr.gouv.tchap.android.sdk.api.services.threepidplatformdiscover.ThreePidPlatformDiscoverService
+import fr.gouv.tchap.android.sdk.internal.services.threepidplatformdiscover.model.Platform
 import okhttp3.OkHttpClient
 import org.matrix.android.sdk.api.session.identity.ThreePid
 import org.matrix.android.sdk.api.session.identity.toMedium
@@ -26,17 +26,17 @@ import org.matrix.android.sdk.internal.di.Unauthenticated
 import org.matrix.android.sdk.internal.network.RetrofitFactory
 import javax.inject.Inject
 
-internal class TchapIdentityService @Inject constructor(
+internal class TchapThreePidPlatformDiscoverService @Inject constructor(
         @Unauthenticated
         private val unauthenticatedOkHttpClient: Lazy<OkHttpClient>,
         private val retrofitFactory: RetrofitFactory,
-        private val identityRequestHomeServerTask: IdentityRequestHomeServerTask
-) : IdentityService {
+        private val threePidPlatformDiscoverTask: ThreePidPlatformDiscoverTask
+) : ThreePidPlatformDiscoverService {
 
     override suspend fun getPlatform(url: String, threePid: ThreePid): Platform {
-        val api = retrofitFactory.create(unauthenticatedOkHttpClient, url).create(IdentityAPI::class.java)
-        val params = IdentityRequestHomeServerTask.Params(api, threePid.value, threePid.toMedium())
+        val api = retrofitFactory.create(unauthenticatedOkHttpClient, url).create(ThreePidPlatformDiscoverAPI::class.java)
+        val params = ThreePidPlatformDiscoverTask.Params(api, threePid.value, threePid.toMedium())
 
-        return identityRequestHomeServerTask.execute(params)
+        return threePidPlatformDiscoverTask.execute(params)
     }
 }
