+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/auth/AuthModule.kt
@@ -20,10 +20,10 @@ import android.content.Context
 import dagger.Binds
 import dagger.Module
 import dagger.Provides
-import fr.gouv.tchap.android.sdk.api.session.identity.IdentityService
-import fr.gouv.tchap.android.sdk.internal.session.identity.DefaultIdentityRequestHomeServerTask
-import fr.gouv.tchap.android.sdk.internal.session.identity.IdentityRequestHomeServerTask
-import fr.gouv.tchap.android.sdk.internal.session.identity.TchapIdentityService
+import fr.gouv.tchap.android.sdk.api.services.threepidplatformdiscover.ThreePidPlatformDiscoverService
+import fr.gouv.tchap.android.sdk.internal.services.threepidplatformdiscover.TchapThreePidPlatformDiscoverService
+import fr.gouv.tchap.android.sdk.internal.services.threepidplatformdiscover.TchapThreePidPlatformDiscoverTask
+import fr.gouv.tchap.android.sdk.internal.services.threepidplatformdiscover.ThreePidPlatformDiscoverTask
 import org.matrix.android.sdk.api.auth.AuthenticationService
 import org.matrix.android.sdk.api.legacy.LegacySessionImporter
 import org.matrix.android.sdk.internal.auth.db.AuthRealmMigration
@@ -81,7 +81,7 @@ internal abstract class AuthModule {
     abstract fun bindAuthenticationService(service: DefaultAuthenticationService): AuthenticationService
 
     @Binds
-    abstract fun bindIdentityService(service: TchapIdentityService): IdentityService
+    abstract fun bindThreePidPlatformDiscoverService(service: TchapThreePidPlatformDiscoverService): ThreePidPlatformDiscoverService
 
     @Binds
     abstract fun bindSessionCreator(creator: DefaultSessionCreator): SessionCreator
@@ -90,7 +90,7 @@ internal abstract class AuthModule {
     abstract fun bindDirectLoginTask(task: DefaultDirectLoginTask): DirectLoginTask
 
     @Binds
-    abstract fun bindIdentityRequestHomeServerTask(task: DefaultIdentityRequestHomeServerTask): IdentityRequestHomeServerTask
+    abstract fun bindThreePidPlatformDiscoverTask(task: TchapThreePidPlatformDiscoverTask): ThreePidPlatformDiscoverTask
 
     @Binds
     abstract fun bindHomeServerHistoryService(service: DefaultHomeServerHistoryService): HomeServerHistoryService
